Some clean up. Linux version working again! Mac OSX compiling!! muParser updated. FreeImage error handled. Director log messages colorized.
git-svn-id: https://moldeo.svn.sourceforge.net/svnroot/moldeo/trunk@105 c963c68c-f84c-0410-a3d6-bef16c67260e

+++ src/moldeodirector/moValueCtrl.cpp
@@ -411,7 +411,7 @@ moValueCtrl::OnSpinUpdated( wxSpinEvent& event) {
 
     moValue& rValue( m_ValueDescriptor.GetValue() );
 
-    switch( m_ValueDescriptor.GetParamDescriptor().GetParamDefinition().GetType()) {
+    switch( (int)m_ValueDescriptor.GetParamDescriptor().GetParamDefinition().GetType()) {
 
         case MO_PARAM_NUMERIC:
             if (m_pSpinCtrl) rValue.GetSubValue(0).SetInt( m_pSpinCtrl->GetValue() );
@@ -544,7 +544,7 @@ moValueCtrl::OnComboUpdated( wxCommandEvent& event) {
     if (m_bInit) {
 
         moValue& rValue( m_ValueDescriptor.GetValue() );
-        switch( m_ValueDescriptor.GetParamDescriptor().GetParamDefinition().GetType()) {
+        switch( (int)m_ValueDescriptor.GetParamDescriptor().GetParamDefinition().GetType()) {
             case MO_PARAM_BLENDING:
                 if (m_pBlendingCtrl) {
                     int i,isel = 0;
